Allow Caching of OWS Responses
At the moment, approximately none (0.44% in the last 30 days) of our
OWS traffic is being cached by CloudFront. This is a problem for two
reasons. 1: poor performance for OWS Users. 2: Increased load on OWS
Application Servers.

This change allows GetCapabilities requests to be cached for 1 hour,
and allows GetMaps/GetTile requests to be cached for 1 hour.

Previously, the only requests being cached are those for GetMap/GetTile
requests that combine 4 or more datasets.